Some updates for linux/gcc and when mysql is not configured

+++ src/bindable.h
@@ -14,7 +14,7 @@ namespace arg3
         {
         public:
             virtual bindable &bind(size_t index, int value) = 0;
-            virtual bindable &bind(size_t index, int64_t value) = 0;
+            virtual bindable &bind(size_t index, long long value) = 0;
             virtual bindable &bind(size_t index, double value) = 0;
             virtual bindable &bind(size_t index, const std::string &value, int len = -1) = 0;
             virtual bindable &bind(size_t index, const sql_blob &value) = 0;
