+++ src/sqlite3_statement.h
@@ -35,7 +35,7 @@ namespace arg3
             string last_error();
             long long last_insert_id();
             sqlite3_statement &bind(size_t index, int value);
-            sqlite3_statement &bind(size_t index, int64_t value);
+            sqlite3_statement &bind(size_t index, long long value);
             sqlite3_statement &bind(size_t index, double value);
             sqlite3_statement &bind(size_t index, const std::string &value, int len = -1);
             sqlite3_statement &bind(size_t index, const sql_blob &value);
